Custom naming policy: retarget net10.0, consolidate four projects to two, NUnit 4
- Both surviving projects move from net7 to net10.0.
- The .NET8 sample and its four tests merge into CustomNamingPolicy and Test;
  both .NET8 projects are deleted and the solution loses their entries. The
  split only existed because the .NET 8 policies would not compile on net7.
  CustomNamingPolicy.NET8.Test had no ProjectReference, so nothing depended on
  the removed projects.
- Test packages to current stable: NUnit 4.6.1, NUnit3TestAdapter 6.3.0,
  NUnit.Analyzers 4.14.0, Microsoft.NET.Test.Sdk 18.9.0, coverlet.collector
  10.0.1. The seven assertions move to Assert.That(..., Is.EqualTo(...)).
- Program.cs converts to top-level statements, which also removes a class that
  shared its name with its namespace.
- NodeSeparatorPolicy uses ArgumentNullException.ThrowIfNull and
  CamelCasePolicy uses a range indexer; both land in the article's snippets.

Build: 0 warnings, 0 errors. Tests: 7/7 passing. SDK 10.0.302, runtime 10.0.10.
