--- json-csharp/CustomNamingPolicy/CustomNamingPolicy.NET8.Test/GlobalUsings.cs
@@ -1,1 +1,0 @@
-global using NUnit.Framework;
