+++ json-csharp/CustomNamingPolicy/CustomNamingPolicy.sln
@@ -7,10 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomNamingPolicy", "Custo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Test", "Test\Test.csproj", "{F0A97070-D698-4309-BD2A-19C1EF410EAA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomNamingPolicy.NET8", "CustomNamingPolicy.NET8\CustomNamingPolicy.NET8.csproj", "{EA2B17EB-E1B1-4FF5-A64D-379DF523F1FE}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomNamingPolicy.NET8.Test", "CustomNamingPolicy.NET8.Test\CustomNamingPolicy.NET8.Test.csproj", "{D9E5E15E-E8E7-470D-B634-D46E786F6889}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -25,14 +21,6 @@ Global
 		{F0A97070-D698-4309-BD2A-19C1EF410EAA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F0A97070-D698-4309-BD2A-19C1EF410EAA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F0A97070-D698-4309-BD2A-19C1EF410EAA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EA2B17EB-E1B1-4FF5-A64D-379DF523F1FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA2B17EB-E1B1-4FF5-A64D-379DF523F1FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA2B17EB-E1B1-4FF5-A64D-379DF523F1FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA2B17EB-E1B1-4FF5-A64D-379DF523F1FE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D9E5E15E-E8E7-470D-B634-D46E786F6889}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D9E5E15E-E8E7-470D-B634-D46E786F6889}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D9E5E15E-E8E7-470D-B634-D46E786F6889}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D9E5E15E-E8E7-470D-B634-D46E786F6889}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
